Introduces Subscriptions component in the app to make the page component cleaner
