+++ apps/example-orchestrator-ui/pages/subscriptions.tsx
@@ -1,173 +1,11 @@
 import React from 'react';
-import {
-    getStatusBadgeColor,
-    Subscriptions,
-    TableColumns,
-} from '@orchestrator-ui/orchestrator-ui-components';
 import NoSSR from 'react-no-ssr';
-import { graphql } from '../__generated__';
-import {
-    MyBaseSubscriptionEdge,
-    PythiaSortOrder,
-    SubscriptionGridQuery,
-    SubscriptionGridQueryVariables,
-    SubscriptionsSort,
-} from '../__generated__/graphql';
-import { TypedDocumentNode } from '@graphql-typed-document-node/core';
-import { EuiBadge } from '@elastic/eui';
-import Link from 'next/link';
-
-// Graphql query, can be moved to separate file / constants
-const GET_SUBSCRIPTIONS_PAGINATED: TypedDocumentNode<
-    SubscriptionGridQuery,
-    SubscriptionGridQueryVariables
-> = graphql(`
-    query SubscriptionGrid(
-        $first: Int!
-        $after: Int!
-        $sortBy: [SubscriptionsSort!]
-    ) {
-        subscriptions(first: $first, after: $after, sortBy: $sortBy) {
-            edges {
-                node {
-                    note
-                    name
-                    startDate
-                    endDate
-                    tag
-                    vlanRange
-                    description
-                    product {
-                        name
-                        type
-                        tag
-                    }
-                    organisation {
-                        abbreviation
-                        name
-                    }
-                    insync
-                    status
-                    subscriptionId
-                }
-            }
-        }
-    }
-`);
-const defaultSortOrder: SubscriptionsSort[] = [
-    { field: 'startDate', order: PythiaSortOrder.Desc },
-];
-const variables: SubscriptionGridQueryVariables = {
-    first: 20,
-    after: 20,
-    sortBy: defaultSortOrder,
-};
-////////////////// Till here /////////////////////////////
-
-// The fields that will end up in the table (columns)
-type Subscription = {
-    subscriptionId: string;
-    description: string;
-    status: string;
-    insync: boolean;
-    startDate: string;
-    productName: string;
-    organisationAbbreviation: string;
-};
+import { Subscriptions } from '../components/Subscriptions/Subscriptions';
 
 export default function SubscriptionsPage() {
-    // Config per column
-    const tableColumnConfig: TableColumns<Subscription> = {
-        description: {
-            displayAsText: 'Description',
-            initialWidth: 400,
-            renderCell: (cellValue) => <h1>{cellValue}</h1>,
-        },
-        insync: {
-            renderCell: (cellValue) => (
-                <EuiBadge
-                    color={cellValue ? 'success' : 'danger'}
-                    isDisabled={false}
-                >
-                    {cellValue.toString()}
-                </EuiBadge>
-            ),
-        },
-        organisationAbbreviation: {
-            displayAsText: 'Customer',
-            initialWidth: 200,
-        },
-        productName: {
-            displayAsText: 'Product',
-            initialWidth: 250,
-        },
-        startDate: {
-            renderCell: (cellValue) =>
-                cellValue
-                    ? new Date(parseInt(cellValue) * 1000).toLocaleString(
-                          'nl-NL',
-                      )
-                    : '',
-        },
-        status: {
-            displayAsText: 'Status',
-            renderCell: (cellValue) => (
-                <EuiBadge
-                    color={getStatusBadgeColor(cellValue)}
-                    isDisabled={false}
-                >
-                    {cellValue}
-                </EuiBadge>
-            ),
-        },
-        subscriptionId: {
-            displayAsText: 'ID',
-            renderCell: (cellValue) => (
-                <Link href={`/subscriptions/${cellValue}`}>
-                    {cellValue.slice(0, 8)}
-                </Link>
-            ),
-        },
-    };
-
-    // You might not want to convert, flatten or ignore data from the api
-    // This mapper converts from the api response (SubscriptionGridQuery)
-    // to the self defined type above (Subscription)
-    const mapApiResponseToTableData = (
-        graphqlResponse: SubscriptionGridQuery,
-    ): Subscription[] =>
-        graphqlResponse.subscriptions.edges.map(
-            (baseSubscription: MyBaseSubscriptionEdge): Subscription => {
-                const {
-                    description,
-                    insync,
-                    organisation,
-                    product,
-                    startDate,
-                    status,
-                    subscriptionId,
-                } = baseSubscription.node;
-
-                return {
-                    description,
-                    insync,
-                    organisationAbbreviation: organisation.abbreviation,
-                    productName: product.name,
-                    startDate,
-                    status,
-                    subscriptionId,
-                };
-            },
-        );
-
     return (
         <NoSSR>
-            <Subscriptions
-                tableColumns={tableColumnConfig}
-                query={GET_SUBSCRIPTIONS_PAGINATED}
-                queryVars={variables}
-                mapApiResponseToTableData={mapApiResponseToTableData}
-            ></Subscriptions>
+            <Subscriptions />
         </NoSSR>
     );
 }
